@@ -1,14 +1,12 @@
 package com.project.safewheels;
 
-import android.content.Context;
-import android.net.Uri;
+import androidx.fragment.app.*;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.annotation.Nullable;
 
 
 public class SecurityFragment extends Fragment {
